+++ resources/lang/en/texts.php
@@ -1,20 +1,24 @@
 <?php
 
 return [
-    'alert_accept' => 'Agreed',
-    'alert_settings' => 'Settings',
-    'alert_text' => 'This website uses cookies to improve your browsing experience. By clicking on ‘Agree’, you accept the use of these cookies.',
+    'alert_accept' => 'Accept all cookies',
+    'alert_essentials_only' => 'Accept only necessary cookies',
+    'alert_settings' => 'Adjust your preferences',
+    'alert_text' => 'By choosing "Accept all cookies" you agree to the use of cookies to help us provide you with a better user experience and to analyse website usage. By clicking "Adjust your preferences" you can choose which cookies to allow. Only the essential cookies are necessary for the proper functioning of our website and cannot be refused',
     'alert_title' => 'This website uses cookies',
+    'alert_essential_only' => 'Accept only necessary cookies',
     'setting_analytics' => 'Analytical cookies',
     'setting_analytics_text' => 'allow us to analyse website use and to improve the visitor\'s experience.',
-    'setting_essential' => 'Functional cookies',
-    'setting_essential_text' => 'are essential cookies that ensure that the website functions properly and that your preferences (e.g. language, region) are saved.',
+    'setting_essential' => 'Essential cookies',
+    'setting_essential_text' => 'are necessary for technical reasons. Without them, this website may not function properly.',
+    'setting_functional' => 'Functional cookies',
+    'setting_functional_text' => 'are necessary for specific functionality on the website. Without them, some features may be disabled.',
     'setting_marketing' => 'Marketing cookies',
     'setting_marketing_text' => 'allow us to personalise your experience and to send you relevant content and offers, on this website and other websites.',
     'settings_accept_all' => 'Accept all cookies',
     'settings_cancel' => 'Cancel',
     'settings_close' => 'Close',
-    'settings_save' => 'Save',
+    'settings_save' => 'Save my selection',
     'settings_title' => 'Cookie settings',
-    'settings_text' => 'Our website keeps three levels of cookies. You can adjust your preferences at any time. If you want more information about what cookies are and which cookies we collect, please read our <a href=":policyUrl">cookie policy</a>.',
+    'settings_text' => 'Our website stores four types of cookies. At any time you can choose which cookies you accept and which you refuse. You can read more about what cookies are and what types of cookies we store in our  <a href=":policyUrl" target="_blank" class="underline hover:no-underline">Cookie Policy</a>.'
 ];