+++ resources/lang/nl/texts.php
@@ -1,20 +1,24 @@
 <?php
 
 return [
-    'alert_accept' => 'Akkoord',
-    'alert_settings' => 'Cookievoorkeuren',
-    'alert_text' => 'We gebruiken cookies om je een betere bezoekerservaring te bieden. Klik op akkoord om alle cookies te accepteren of ga naar cookieinstellingen om je voorkeuren te beheren.',
+    'alert_accept' => 'Alle cookies aanvaarden',
+    'alert_essentials_only' => 'Enkel essentiële cookies aanvaarden',
+    'alert_settings' => 'Je voorkeuren aanpassen',
+    'alert_text' => 'Door op "Alle cookies aanvaarden" te klikken ga je akkoord met het opslaan van onze cookies op je computer. Zo kunnen we een betere gebruikservaring bieden, en het laat ons toe om het gebruik van de website te analyseren. Door op "Je voorkeuren aanpassen" te klikken kan je kiezen welke cookies je toelaat. Enkel de essentiële cookies zijn noodzakelijk voor de goede werking van de website en kan je niet weigeren.',
     'alert_title' => 'Deze website gebruikt cookies',
+    'alert_essential_only' => 'Enkel essentiële cookies aanvaarden',
     'setting_analytics' => 'Analytische cookies',
     'setting_analytics_text' => 'laten ons toe om het gebruik van de website te analyseren en de bezoekerservaring te verbeteren.',
-    'setting_essential' => 'Functionele cookies',
-    'setting_essential_text' => 'zijn noodzakelijke cookies die ervoor zorgen dat de website goed functioneert en dat je voorkeuren (bv. taal, regio) goed worden opgeslagen.',
+    'setting_essential' => 'Essentiële cookies',
+    'setting_essential_text' => 'zijn om technische redenen noodzakelijk voor de goede werking. Zonder deze cookies zou het kunnen dat de website niet naar behoren functioneert.',
+    'setting_functional' => 'Functionele cookies',
+    'setting_functional_text' => 'zijn noodzakelijk voor bepaalde functies op de website. Zonder deze cookies kunnen sommige onderdelen onbruikbaar worden.',
     'setting_marketing' => 'Marketingcookies',
     'setting_marketing_text' => 'hebben als doel je ervaring te personaliseren en je relevante inhoud en aanbiedingen te sturen op deze en andere websites.',
     'settings_accept_all' => 'Alle cookies aanvaarden',
     'settings_cancel' => 'Annuleren',
     'settings_close' => 'Sluiten',
-    'settings_save' => 'Opslaan',
+    'settings_save' => 'Mijn keuze bewaren',
     'settings_title' => 'Cookievoorkeuren',
-    'settings_text' => 'Onze website houdt drie niveaus van cookies bij. Je kan zelf je voorkeuren op elk moment aanpassen. Meer informatie over wat cookies zijn en welke cookies we verzamelen kan je lezen in ons <a href=":policyUrl">cookiebeleid</a>.',
+    'settings_text' => 'Onze website bewaart vier soorten cookies. Je kan zelf op elk moment kiezen welke je aanvaardt en welke niet. Meer informatie over wat cookies zijn en welke types cookies we verzamelen kan je lezen in ons <a href=":policyUrl" target="_blank" >cookiebeleid</a>.'
 ];