+++ resources/views/index.blade.php
@@ -17,15 +17,18 @@
             @lang('cookie-consent::texts.alert_title')
         </h2>
         <p id="lcc-modal-alert-desc" class="lcc-text">
-            @lang('cookie-consent::texts.alert_text')
+            {!! trans('cookie-consent::texts.alert_text') !!}
         </p>
     </div>
     <div class="lcc-modal__actions">
-        <button type="button" class="lcc-button lcc-button--link js-lcc-settings-toggle">
-            @lang('cookie-consent::texts.alert_settings')
-        </button>
         <button type="button" class="lcc-button js-lcc-accept">
             @lang('cookie-consent::texts.alert_accept')
+        </button>
+        <button type="button" class="lcc-button lcc-button--link js-lcc-essentials">
+            @lang('cookie-consent::texts.alert_essential_only')
+        </button>
+        <button type="button" class="lcc-button lcc-button--link js-lcc-settings-toggle">
+            @lang('cookie-consent::texts.alert_settings')
         </button>
     </div>
 </div>
@@ -43,7 +46,7 @@
                 @lang('cookie-consent::texts.settings_title')
             </h2>
             <p id="lcc-modal-settings-desc" class="lcc-text">
-                @lang('cookie-consent::texts.settings_text', [ 'policyUrl' => config("cookie-consent.policy_url_$locale")])
+                {!! trans('cookie-consent::texts.settings_text', [ 'policyUrl' => config("cookie-consent.policy_url_$locale")]) !!}
             </p>
             <div class="lcc-modal__section lcc-u-text-center">
                 <button type="button" class="lcc-button js-lcc-accept">
@@ -57,6 +60,15 @@
                 </label>
                 <p class="lcc-text">
                     @lang('cookie-consent::texts.setting_essential_text')
+                </p>
+            </div>
+            <div class="lcc-modal__section">
+                <label for="lcc-checkbox-funtcional" class="lcc-label">
+                    <input type="checkbox" id="lcc-checkbox-funtcional" disabled="disabled" checked="checked">
+                    <span>@lang('cookie-consent::texts.setting_functional')</span>
+                </label>
+                <p class="lcc-text">
+                    @lang('cookie-consent::texts.setting_functional_text')
                 </p>
             </div>
             <div class="lcc-modal__section">
@@ -80,9 +92,6 @@
         </div>
     </div>
     <div class="lcc-modal__actions">
-        <button type="button" class="lcc-button lcc-button--link js-lcc-settings-toggle">
-            @lang('cookie-consent::texts.settings_cancel')
-        </button>
         <button type="button" class="lcc-button js-lcc-settings-save">
             @lang('cookie-consent::texts.settings_save')
         </button>